bit manipulation added

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -7,7 +7,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	size_t c = 0;
+	unsigned long int c = 0;
 	unsigned long int x_result = n ^ m;
 
 	while (x_result > 0)
